sched: EAS: energy model change
Change model to describe SMP for most of the power curve. The siver/gold
clusters are very close and may only differ by measurement error for
most of the curve. Several workloads (including video playback)
have better power characteristics if we go ahead and make the curves
identical so it looks like an SMP system for frequencies above the
min and below the silver max.

Change-Id: I16d247cad295101b52eb70d82a8f3977e819db6a